force int for k index

+++ Containers/UnstructuredMeshTools.py
@@ -1824,7 +1824,7 @@ def GetValueAtPosLinearSymplecticMesh(field,mesh,constantRectilinearMesh):
                     jmin, jmax = int(math.floor((localBoundingMin[1]-origin[1])/spacing[1])),int(math.floor((localBoundingMax[1]-origin[1])/spacing[1])+1)
                     
                     if mesh.GetDimensionality()>2:
-                        kmin, kmax = math.floor((localBoundingMin[2]-origin[2])/spacing[2]),math.floor((localBoundingMax[2]-origin[2])/spacing[2])+1
+                        kmin, kmax = int(math.floor((localBoundingMin[2]-origin[2])/spacing[2])),int(math.floor((localBoundingMax[2]-origin[2])/spacing[2])+1)
                         
                     """imin, imax = math.floor((localBoundingMin[0])/spacing[0]),math.floor((localBoundingMax[0])/spacing[0])+1
                     jmin, jmax = math.floor((localBoundingMin[1])/spacing[1]),math.floor((localBoundingMax[1])/spacing[1])+1
